backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pytest
black
flake8
